+++ sdk/core/src/models/MoESupportedAttributes.ts
@@ -1,0 +1,6 @@
+export type MoESupportedAttributes =
+    | string
+    | number
+    | boolean
+    | Array<any>
+    | { [k: string]: any }