+++ sdk/core/src/utils/MoEJsonBuilder.ts
@@ -123,7 +123,7 @@ export function getAppStatusJson(appStatus: String, appId: String) {
   return JSON.stringify(json);
 }
 
-export function getUserAttributeJson(name: String, value: String | Number | Boolean | Array<String> | Array<Number>, type: String, appId: String) {
+export function getUserAttributeJson(name: String, value: MoESupportedAttributes, type: String, appId: String) {
   var json: { [k: string]: any } = {
     accountMeta: {
       appId: appId